# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.AttachmentBatchDeleteFailedItem:
    properties:
      id:
        type: integer
      message:
        type: string
    type: object
  dto.AttachmentBatchDeleteInput:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
  dto.AttachmentBatchDeleteResult:
    properties:
      failed:
        items:
          $ref: '#/definitions/dto.AttachmentBatchDeleteFailedItem'
        type: array
      success_ids:
        items:
          type: integer
        type: array
    type: object
  dto.AttachmentDTO:
    properties:
      biz_type:
        type: integer
      created_at:
        type: string
      file_key:
        type: string
      file_size:
        type: integer
      id:
        type: integer
      storage_type:
        type: string
      url:
        type: string
      user_id:
        type: integer
    type: object
  dto.AttachmentListResult:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.AttachmentDTO'
        type: array
      total:
        type: integer
    type: object
...
  dto.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 这是一个轻量级的说说服务
  title: EasyDrop API
  version: "1.0"
paths:
  /admin/attachments:
    get:
      description: 分页查询附件列表，支持按用户、业务类型、上传时间区间过滤
      parameters:
      - description: 附件ID
        in: query
        name: id
        type: integer
      - description: 用户ID
        in: query
        name: user_id
        type: integer
      - description: 附件业务类型
        in: query
        name: biz_type
        type: integer
      - description: 上传时间起点（Unix秒）
        in: query
        name: created_from
        type: integer
      - description: 上传时间终点（Unix秒）
        in: query
        name: created_to
        type: integer
      - description: 页码，从 1 开始
        in: query
        name: page
        type: integer
      - description: 每页条数，最大 100
        in: query
        name: size
        type: integer
      - description: 排序，如 created_at_desc
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AttachmentListResult'
        "400":
          description: 参数校验失败
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: 未登录或登录失效
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: 服务内部错误
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 管理端查询附件列表
      tags:
      - attachment-admin
  /admin/attachments/{id}:
    delete:
      description: 按附件 ID 删除任意用户附件
      parameters:
      - description: 附件ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "400":
          description: 参数校验失败
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: 未登录或登录失效
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: 附件不存在
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: 服务内部错误
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 管理端删除附件
      tags:
      - attachment-admin
  /admin/attachments/batch-delete:
    post:
      consumes:
      - application/json
      description: 批量删除附件，返回成功与失败明细
      parameters:
      - description: 批量删除请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AttachmentBatchDeleteInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AttachmentBatchDeleteResult'
        "400":
          description: 参数校验失败
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: 未登录或登录失效
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: 服务内部错误
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 管理端批量删除附件
      tags:
      - attachment-admin
...
  /attachments:
    get:
      description: 分页查询当前用户附件列表
      parameters:
      - description: 附件ID
        in: query
        name: id
        type: integer
      - description: 附件业务类型
        in: query
        name: biz_type
        type: integer
      - description: 页码，从 1 开始
        in: query
        name: page
        type: integer
      - description: 每页条数，最大 100
        in: query
        name: size
        type: integer
      - description: 排序，如 created_at desc
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AttachmentListResult'
        "400":
          description: 参数校验失败
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: 未登录或登录失效
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: 服务内部错误
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 查询附件列表
      tags:
      - attachment
    post:
      consumes:
      - multipart/form-data
      description: 以 multipart/form-data 方式上传附件并返回附件信息
      parameters:
      - description: 附件文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.AttachmentDTO'
        "400":
          description: 参数校验失败
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: 未登录或登录失效
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: 存储配额不足
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: 用户不存在
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: 服务内部错误
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 上传附件
      tags:
      - attachment
  /attachments/{id}:
    delete:
      description: 删除当前用户的附件
      parameters:
      - description: 附件ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "400":
          description: 参数校验失败
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: 未登录或登录失效
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: 附件不存在
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: 服务内部错误
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 删除附件
      tags:
      - attachment
    get:
      description: 根据附件 ID 查询当前用户的附件详情
      parameters:
      - description: 附件ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AttachmentDTO'
        "400":
          description: 参数校验失败
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: 未登录或登录失效
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: 附件不存在
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: 服务内部错误
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 查询附件详情
      tags:
      - attachment
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"